% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{default_config_dir}
\alias{default_config_dir}
\title{Get the default Snowflake configuration directory}
\usage{
default_config_dir(os = NULL)
}
\arguments{
\item{os}{Operating system identifier; one of "win", "mac", or "unix".
If NULL (the default), the value is determined automatically.}
}
\value{
Path to the default Snowflake configuration directory
}
\description{
Get the default Snowflake configuration directory
}
\keyword{internal}
