% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{has_a_default_connection}
\alias{has_a_default_connection}
\title{Reports whether a default connection is available}
\usage{
has_a_default_connection(...)
}
\arguments{
\item{...}{arguments passed to \code{\link[=snowflake_connection]{snowflake_connection()}}}
}
\description{
Reports whether a default connection is available
}
\examples{
has_a_default_connection()
}
