% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowflakeauth-package.R
\docType{package}
\name{snowflakeauth-package}
\alias{snowflakeauth}
\alias{snowflakeauth-package}
\title{snowflakeauth: Authentication Helpers for 'Snowflake'}
\description{
Authentication helpers for 'Snowflake'. It provides compatibility with authentication approaches supported by the 'Snowflake Connector for Python' \url{https://pypi.org/project/snowflake-connector-python} and the 'Snowflake CLI' \url{https://pypi.org/project/snowflake-cli}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://posit-dev.github.io/snowflakeauth/}
  \item \url{https://github.com/posit-dev/snowflakeauth}
  \item Report bugs at \url{https://github.com/posit-dev/snowflakeauth/issues}
}

}
\author{
\strong{Maintainer}: Aaron Jacobs \email{aaron.jacobs@posit.co}

Authors:
\itemize{
  \item E. David Aja \email{david@posit.co}
}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
