% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-credentials.R
\name{snowflake_credentials}
\alias{snowflake_credentials}
\title{Get credentials for a Snowflake connection}
\usage{
snowflake_credentials(params, role = NULL, spcs_endpoint = NULL, ...)
}
\value{
A list of HTTP headers.
}
\description{
Get credentials for a Snowflake connection
}
