% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowflakeauth-package.R
\docType{package}
\name{snowflakeauth-package}
\alias{snowflakeauth}
\alias{snowflakeauth-package}
\title{snowflakeauth: Support for authenticating with the 'Snowflake' platform}
\description{
TODO
}
\author{
\strong{Maintainer}: Aaron Jacobs \email{aaron.jacobs@posit.co}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
