% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowflakeauth-package.R
\docType{package}
\name{snowflakeauth-package}
\alias{snowflakeauth}
\alias{snowflakeauth-package}
\title{snowflakeauth: Authentication Helpers for 'Snowflake'}
\description{
A toolkit for authenticating with 'Snowflake'. It provides compatibility with authentication apporaches supported by the 'Snowflake Connector for Python' and the 'Snowflake CLI'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://posit-dev.github.io/snowflakeauth/}
}

}
\author{
\strong{Maintainer}: E. David Aja \email{david@posit.co}

Authors:
\itemize{
  \item Aaron Jacobs \email{aaron.jacobs@posit.co}
}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
